+++ Runtime/CustomAttributes/NonCustomClass.cs
@@ -1,0 +1,11 @@
+using System;
+using UnityEngine;
+
+namespace CustomAttributes
+{
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct)]
+    public class NonCustomClass : PropertyAttribute
+    {
+    
+    }
+}